% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\docType{data}
\name{Activation.Identity}
\alias{Activation.Identity}
\title{Identity Activation object}
\format{
An object of class \code{Activation} of length 3.
}
\usage{
Activation.Identity
}
\description{
Identity Activation object
}
\examples{
Dense(3, 4, Activation.Identity) # a 3 -> 4 Dense network with Identity activation.
}
\keyword{datasets}
