% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\name{Activation}
\alias{Activation}
\title{Activation class constructor}
\usage{
Activation(f, df, name)
}
\arguments{
\item{f}{A function to compute f(z).}

\item{df}{A function to compute df/dz, input parameters are z and a.}

\item{name}{name of the activation.}
}
\value{
An Activation object.
}
\description{
Activation class constructor
}
\examples{
f <- function(z) z
df <- function(z, a) rep(1, length(z))
identity <- Activation(f, df, "identity")
identity
}
