% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\docType{data}
\name{Activation.ReLU}
\alias{Activation.ReLU}
\title{ReLU Activation object}
\format{An object of class \code{Activation} of length 3.}
\usage{
Activation.ReLU
}
\description{
ReLU Activation object
}
\examples{
z <- seq(from=-5, to=5, length=201)
a <- Activation.ReLU$f(z)
plot(a ~ z, main='ReLU')
plot(Activation.ReLU$df(a) ~ a, main='ReLU df')
}
\keyword{datasets}
