% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\docType{data}
\name{Activation.Sigmoid}
\alias{Activation.Sigmoid}
\title{Sigmoid Activation object}
\format{An object of class \code{Activation} of length 3.}
\usage{
Activation.Sigmoid
}
\description{
Sigmoid Activation object
}
\examples{
z <- seq(from=-5, to=5, length=201)
plot(Activation.Sigmoid$f(z) ~ z)
}
\keyword{datasets}
