% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{Backward}
\alias{Backward}
\title{Generic Backward function}
\usage{
Backward(object, errorOut)
}
\arguments{
\item{object}{Object that supports Backward function}

\item{errorOut}{partial Cost / partial out}
}
\value{
a layer object with done = "Backward"
}
\description{
Generic Backward function
}
