% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mnist.R
\name{BuildBatch}
\alias{BuildBatch}
\title{BuildBatch is to build a batch "generator" given a dataset.}
\usage{
BuildBatch(data)
}
\arguments{
\item{data}{a list with two required elements ("x" and "y") and one optional
element ("weights"), where data$x is a matrix with dimension nvars x nobs,
data$y is a vector or a matrix. If data$weights exists, its length should
same as the number of cols in data$x.}
}
\description{
BuildBatch is to build a batch "generator" given a dataset.
}
