% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_conv.R
\name{Conv}
\alias{Conv}
\title{A S3 class to represent Convoluation neuro layer.}
\usage{
Conv(
  inputDim,
  kernelDim,
  activation = Activation.ReLU,
  initWeightScale = 0.1,
  initBias = 0.01
)
}
\arguments{
\item{inputDim}{a size 3 vector to indicate the dimension of one observation.}

\item{kernelDim}{a size 4 vector to indicate the dimension of kernel}

\item{activation}{activation to be usaed for the neuro, default ReLU()}
}
\description{
A S3 class to represent Convoluation neuro layer.
}
\examples{
inputDim = c(5, 5, 1)
outputDim = c(5, 5, 2)
kernelDim = c(3, 3, 1, 2)
layer = Conv(inputDim=inputDim, kernelDim=kernelDim)
layer

nObs = 1

x = array(rnorm(prod(inputDim)*nObs), dim=c(inputDim, nObs))
errorOut=array(rnorm(prod(outputDim)*nObs), dim=c(outputDim, nObs))

layer = Forward(layer, x)
layer$a # output

layer = Backward(layer, errorOut)
layer$errorIn
}
