% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_compose.R
\name{DAG}
\alias{DAG}
\title{A S3 class to represent a directed acycle graph (DAG) of a list of}
\usage{
DAG(layers, edges)
}
\arguments{
\item{layers}{a list of layers}

\item{edges}{a vector of even intergers. The pair of integers at position
2*i-1 and 2*i represents a directed edge in the DAG. We assumed the number
c(1,2, 2,3) means 1->2, 2->3
c(1,2, 4,5) means 1->2, 4->5
c(1,2, 1,3, 2,3) means 1->2, 1->3, 2->3. Note that if output dim of 1 and 2 is
o1 and o2, respectively, then input dim of 3 is o1 + o2, with first o1 dim from
1, second o2 dim from 2.}
}
\description{
A S3 class to represent a directed acycle graph (DAG) of a list of
}
\examples{
layers = list(Dense(3,5), Dense(5, 8), Dense(8, 7), Dense(8, 6), Softmax(6))
edges = list(c(1,2,3), c(2,4,5))
# net has inputDim of 3, outputDim of 8 + 6
net = DAG(layers, c(1,2, 2,3, 2,4, 4,5))
net
}
