% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_dense.R
\name{Dense}
\alias{Dense}
\title{A S3 class to represent Dense (full-connect) neuro layer.}
\usage{
Dense(
  inputDim,
  outputDim,
  activation = Activation.ReLU,
  initWeightScale = 0.01,
  initBias = 0
)
}
\arguments{
\item{inputDim}{integer, refers to the dimension of one observation.}

\item{outputDim}{integer, refers to the dimension of one output.}

\item{activation}{activation to be used for the neuro, default ReLU()}
}
\description{
A S3 class to represent Dense (full-connect) neuro layer.
}
\examples{
layer = Dense(inputDim=3, outputDim=5)
layer
names(layer)
layer$W # weight parameters
layer$b # bias parameters

x = array(rnorm(3*2), dim=c(3,2))
errorOut=array(rnorm(10), dim=c(5,2))

layer = Forward(layer, x)
layer$a # output

layer = Backward(layer, errorOut)
layer$errorIn

layer = UpdateParameters(layer, 0.1)
layer$b
}
