% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_dropout.R
\name{Dropout}
\alias{Dropout}
\title{A S3 class to represent Dropout layer.}
\usage{
Dropout(inputDim, keepProb = 0.5)
}
\arguments{
\item{inputDim}{an integer or vector to indicate dimension of one input.}

\item{keepProb}{probability to keep a neuron.}
}
\description{
A S3 class to represent Dropout layer.
}
\examples{
layer = Dropout(4, keepProb=0.5)
layer

nObs = 2
x = array(rnorm(4*nObs), c(4, nObs))
x

layer = Forward(layer, x) # Forward for training
layer$a

layer = Forward(layer, x, stage="inference") # Forward for inference
layer$a
}
