% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{Forward}
\alias{Forward}
\title{Generic Forward function}
\usage{
Forward(object, x, ...)
}
\arguments{
\item{object}{Object that supports Forward function}

\item{x}{input data, always assume there are multiple observations in x}

\item{...}{other optional parameter for specail layer like Dropout}
}
\value{
a layer object with done = "Forward"
}
\description{
Generic Forward function
}
