% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{InitParameters.default}
\alias{InitParameters.default}
\title{Default implementation of InitParameters generic function}
\usage{
\method{InitParameters}{default}(object, ...)
}
\description{
Default implementation of InitParameters generic function
}
