% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{InitParameters}
\alias{InitParameters}
\title{Generic InitParameters function}
\usage{
InitParameters(object, ...)
}
\arguments{
\item{object}{Object that supports InitParameters function}

\item{...}{other arguments}
}
\description{
Generic InitParameters function
}
