% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mnist.R
\name{LoadMnist}
\alias{LoadMnist}
\title{Load the MNIST digit recognition dataset into R}
\usage{
LoadMnist(dataDir, byrow = TRUE)
}
\arguments{
\item{dataDir}{directory where MNIST data is in.}
}
\value{
a list(train=train, test=test), where

train$x is 60000x784 matrix, each row is one digit (28x28), if byrow=TRUE,
otherwise, train$x is 784x60000 matrix, each col is one digit(28x28).
train$y is 60000-length vector.

test$x is 10000x784 matrix, test$y is 10000-length vector, if byrow=TRUE.
}
\description{
Load the MNIST digit recognition dataset into R
}
