% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_maxpool.R
\name{MaxPool}
\alias{MaxPool}
\title{A S3 class to represent Maxpool layer.}
\usage{
MaxPool(inputDim, kernelDim)
}
\arguments{
\item{inputDim}{a size 3 vector to indicate the dimension of one observation.}

\item{kernelDim}{a size 2 vector to indicate the dimension of kernel}
}
\description{
A S3 class to represent Maxpool layer.
}
\examples{
inputDim = c(4, 4, 1)
kernelDim = c(2,2)
layer = MaxPool(inputDim=inputDim, kernelDim=kernelDim)
layer

outputDim = layer$outputDim
nObs = 1
x = array(rnorm(prod(inputDim)*nObs), dim=c(inputDim, nObs))
errorOut = array(rnorm(prod(outputDim)*nObs), dim=c(outputDim, nObs))

layer = Forward(layer, x)
layer$a # output

layer = Backward(layer, errorOut)
layer$errorIn
}
