% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{NumParameters.default}
\alias{NumParameters.default}
\title{Default implementation of NumParameters generic function}
\usage{
\method{NumParameters}{default}(object, ...)
}
\description{
Default implementation of NumParameters generic function
}
