% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{NumParameters}
\alias{NumParameters}
\title{Generic NumParameters function}
\usage{
NumParameters(object, ...)
}
\arguments{
\item{object}{Object that supports NumParameters function}

\item{...}{other arguments}
}
\description{
Generic NumParameters function
}
