% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_reshape.R
\name{Reshape}
\alias{Reshape}
\title{A S3 class to represent Reshape layer.}
\usage{
Reshape(inputDim, outputDim)
}
\arguments{
\item{inputDim}{an integer or vector to indicate the dimension of one input observation.}

\item{outputDim}{an integer or vector to indicate the dimension of one output observation}
}
\description{
A S3 class to represent Reshape layer.
}
\examples{
layer = Reshape(9, c(3, 3, 1))
layer

nObs = 1
x = array(rnorm(9), c(9, 1))
layer = Forward(layer, x)
layer$x # input
layer$a # output
}
