% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_compose.R
\name{Sequential}
\alias{Sequential}
\alias{Strip.Sequential}
\alias{Forward.Sequential}
\alias{Backward.Sequential}
\alias{UpdateParameters.Sequential}
\title{A S3 class to represent a "Layer" that contains a sequence of layers.}
\usage{
Sequential(layer1, layer2, ...)

\method{Strip}{Sequential}(object)

\method{Forward}{Sequential}(object, x, ...)

\method{Backward}{Sequential}(object, errorOut)

\method{UpdateParameters}{Sequential}(object, learningRate)
}
\arguments{
\item{layer1}{a Layer}

\item{layer2}{another Layer}

\item{...}{additional layers.}
}
\description{
A S3 class to represent a "Layer" that contains a sequence of layers.
}
\section{Methods (by generic)}{
\itemize{
\item \code{Strip}: strip the net to only keep what is needed for Forward method.

\item \code{Forward}: implement generic Forward function

\item \code{Backward}: implement generic Backward function.

\item \code{UpdateParameters}: implement generic UpdateParameters function.
}}

\examples{

net <- Sequential(Dense(784, 15), Dense(15, 10, Activation.Identity), Softmax)
net
}
