% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_softmax.R
\name{Softmax}
\alias{Softmax}
\title{A S3 class to represent Softmax layer.}
\usage{
Softmax(numClasses)
}
\arguments{
\item{numClasses}{an integer to indicate the number of classes.}
}
\description{
A S3 class to represent Softmax layer.
}
\examples{
layer = Softmax(3)
layer

nObs = 2
x = array(rnorm(3*nObs), c(3, nObs))
layer = Forward(layer, x)
layer$x # input
layer$a # output
}
