% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{Strip}
\alias{Strip}
\title{Generic Strip function}
\usage{
Strip(object)
}
\arguments{
\item{object}{Object that supports Strip function}
}
\value{
a layer object where anything that is not needed for inference are removed.
}
\description{
Generic Strip function
}
