% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{UpdateParameters.default}
\alias{UpdateParameters.default}
\title{Default implementation of UpdateParameters generic function}
\usage{
\method{UpdateParameters}{default}(object, learningRate)
}
\description{
Default implementation of UpdateParameters generic function
}
