% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{UpdateParameters}
\alias{UpdateParameters}
\title{Generic UpdateParameters function}
\usage{
UpdateParameters(object, learningRate)
}
\arguments{
\item{object}{Object that supports UpdateParameters function}

\item{learningRate}{float, between 0 and 1.0}
}
\value{
a layer object with updated parameters (if applicable)
}
\description{
Generic UpdateParameters function
}
